SC-3135: Fix for SC-3694: docker/sdk export assets does not work properly

+++ bin/sdk/assets/baked.sh
@@ -25,25 +25,36 @@ function Assets::export() {
     rm -rf "${projectDockerAssetsTmpDirectory}"
     mkdir -p "${projectDockerAssetsTmpDirectory}"
 
-    echo -e "Preparing assets archives..." >/dev/stderr
+    local command="true"
+    for entrypoint in "${SPRYKER_ENTRYPOINTS[@]}";
+    do
+        command="${command} && \$([ -d '/data/public/${entrypoint}/assets' ] && tar czf '/data${dockerAssetsTmpDirectory}/assets-${entrypoint}-${tag}.tar' -C '/data/public/${entrypoint}/assets' . || true)"
+    done
+
+    echo -e "Preparing assets archives..." > /dev/stderr
 
     docker run --rm \
         -e PROJECT_DIR='/data' \
-        -v "${DEPLOYMENT_DIR}/bin:/data/bin" \
+        -v "${DEPLOYMENT_DIR}/bin:/data/standalone" \
         -v "${projectDockerAssetsTmpDirectory}:/data${dockerAssetsTmpDirectory}" \
+        --entrypoint='' \
         --name="${SPRYKER_DOCKER_PREFIX}_builder_assets" \
         "${SPRYKER_DOCKER_PREFIX}_builder_assets:${SPRYKER_DOCKER_TAG}" \
-        bash -c "/data/bin/standalone/assets/tar-builder.sh ${tag} /data/${dockerAssetsTmpDirectory}"
+        sh -c "${command}" 2>&1
 
-    echo -e "${INFO}File name:              Path:${NC}"
+    echo -e "${INFO}The following assets archives have been prepared${NC}:" > /dev/stderr
 
-    for filePath in "${projectDockerAssetsTmpDirectory}"*; do
-        local fileName="$(basename -- "${filePath}")"
+    for entrypoint in "${SPRYKER_ENTRYPOINTS[@]}";
+    do
+        local fileName="assets-${entrypoint}-${tag}.tar"
+        if [ ! -f "${projectDockerAssetsTmpDirectory}/${fileName}" ]; then
+            continue
+        fi
 
         rm -f "${destinationPath}/${fileName}"
         mv "${projectDockerAssetsTmpDirectory}/${fileName}" "${destinationPath}"
 
-        echo -e "${OK}${fileName}${NC}          ${destinationPath}/${fileName}"
+        echo "${application} ${destinationPath}/${fileName}"
     done
 
     rm -rf "${projectDockerAssetsTmpDirectory}"
